[REFACTOR]: Change import statement
